move pretty into stable_mir
